removed hello world controller; added users controllers and service implementation

+++ src/config/migrations/002_create_user.h
@@ -22,7 +22,9 @@ public:
 			"user",
 			[](auto& table)
 			{
-				table.template column<size_t>("id", {.primary_key=true, .unique=true, .null=false});
+				table.template column<long long int>(
+					"id", {.primary_key=true, .unique=true, .null=false}
+				);
 				table.template column<std::string>("email", {.max_len=320, .null=false});
 				table.template column<std::string>("password_hash", {.max_len=64, .null=false});
 				table.template column<xw::dt::Datetime>("created_at", {.null=false});
